requirements.txt
absl-py==2.1.0
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comtypes==1.4.8
contourpy==1.3.1
cycler==0.12.1
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.55.0
fsspec==2024.10.0
gTTS==2.5.4
idna==3.10
jax==0.4.35
jaxlib==0.4.35
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mediapipe==0.10.18
ml_dtypes==0.5.0
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging==24.2
pillow==11.0.0
protobuf==4.25.5
pycparser==2.22
pygame==2.6.1
pyparsing==3.2.0
pypiwin32==223
python-dateutil==2.9.0.post0
pyttsx3==2.98
pywin32==308
requests==2.32.3
scipy==1.14.1
sentencepiece==0.2.0
setuptools==75.5.0
six==1.16.0
sounddevice==0.5.1
sympy==1.13.1
torch==2.5.1
torchvision==0.20.1
typing_extensions==4.12.2
urllib3==2.2.3
